3-Dependencies-Resolving
You are implementing a package manager for the new hot language -
Java.JS. This is a Java language that compiles to JavaScript.
There are 3rd party packages that can be installed. All packages are
described in all_packages.json.
Your job is to implement a simple command-line tool that reads
dependencies.json file and outputs what else should be installed in
order for everything to work.