2-Word-Game
You are given a rectangular table filled with characters and a word.
Your task is to count the occurrences of the word in the table. The word
can be found horizontally, vertically and across both left to right and
right to left.
Figure out a smart way to input the required data.